--- Mov/Mov/ViewControllers/FavoritesViewController.swift
@@ -1,36 +1,0 @@
-//
-//  FavoritesViewController.swift
-//  Mov
-//
-//  Created by Victor Leal on 18/07/19.
-//  Copyright © 2019 Victor Leal. All rights reserved.
-//
-
-import UIKit
-
-class FavoritesViewController: UIViewController {
-    
-    let screen = FavoritesViewControllerScreen()
-    
-    override func loadView() {
-        self.view = screen
-        self.view.backgroundColor = .blue
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-    
-    override var prefersStatusBarHidden: Bool {
-        return true
-    }
-    
-    
-}
-
-
-
-
-
-
